[CoreBundle] Moving and slugify workspace url registration.
