+++ main/core/Resources/modules/workspace/user/parameters/components/parameters-tab.jsx
@@ -1,9 +1,13 @@
 import React from 'react'
 import {connect} from 'react-redux'
+
+import {generateUrl} from '#/main/core/api/router'
 import {trans} from '#/main/core/translation'
+
 import {PageActions} from '#/main/core/layout/page/components/page-actions.jsx'
 import {FormContainer} from '#/main/core/data/form/containers/form.jsx'
 import {FormPageActionsContainer} from '#/main/core/data/form/containers/page-actions.jsx'
+
 import {select as formSelect} from '#/main/core/data/form/selectors'
 
 const ParametersTabActions = () =>
@@ -15,47 +19,56 @@ const ParametersTabActions = () =>
     />
   </PageActions>
 
-const Parameters = () =>
-  <FormContainer
-    level={3}
-    name="parameters"
-    sections={[{
-      id: 'general',
-      title: trans('general'),
-      primary: true,
-      fields: [
-        {
-          name: 'registration.validation',
-          type: 'boolean',
-          label: trans('registration_validation')
-        }, {
-          name: 'registration.selfRegistration',
-          type: 'boolean',
-          label: trans('public_registration')
-        }, {
-          name: 'registration.selfUnregistration',
-          type: 'boolean',
-          label: trans('public_unregistration')
-        }, {
-          name: 'display.displayable',
-          type: 'boolean',
-          label: trans('displayable_in_workspace_list')
-        },
-        {
-          name: 'restrictions.maxUsers',
-          type: 'number',
-          label: trans('workspace_max_users')
-        }
-      ]
-    }]}
-  />
+//todo: maybe rename form name: parameters is missleading
+const Parameters = props =>
+  <div>
+    <FormContainer
+      level={3}
+      name="parameters"
+      sections={[{
+        id: 'general',
+        title: trans('general'),
+        primary: true,
+        fields: [
+          {
+            name: 'registration.validation',
+            type: 'boolean',
+            label: trans('registration_validation')
+          }, {
+            name: 'registration.selfRegistration',
+            type: 'boolean',
+            label: trans('public_registration')
+          }, {
+            name: 'registration.selfUnregistration',
+            type: 'boolean',
+            label: trans('public_unregistration')
+          }, {
+            name: 'display.displayable',
+            type: 'boolean',
+            label: trans('displayable_in_workspace_list')
+          },
+          {
+            name: 'restrictions.maxUsers',
+            type: 'number',
+            label: trans('workspace_max_users')
+          }
+        ]
+      }]}
+    />
+    <div className="panel panel-body">
+      <h4 class="panel-title">{trans('generate_url')}</h4> <br />
+      <div className="alert alert-info">
+        {generateUrl('claro_workspace_subscription_url_generate', {slug: props.workspace.meta.slug}, true)}
+      </div>
+    </div>
+  </div>
 
 Parameters.propTypes = {
 }
 
 const ParametersTab = connect(
   (state) => ({
-    workspace: formSelect.data(formSelect.form(state, 'workspace'))
+    workspace: formSelect.data(formSelect.form(state, 'parameters'))
   })
 )(Parameters)
